requirements.txt
waymo-open-dataset-tf-2-11-0==1.5.2
jupyter-dash
dask
tensorflow
